+++ CHANGELOG.rst
@@ -16,6 +16,21 @@ Unreleased
 
 *
 
+0.3.4 - 2025-05-23
+******************
+
+Added
+=====
+
+* Bulk unenrollment API.
+* Enrollment audit model that tracks the enrollment state transitions.
+
+Changed
+=======
+
+* The Learning Paths API includes start and end dates for its steps.
+* Return enrollment date in the API instead of a boolean.
+
 0.3.3 - 2025-05-23
 ******************